requirements.txt
pandas==1.5.1
pyodbc==4.0.34
psycopg2==2.9.5
scikit-learn==1.1.2
Flask==2.2.2
Flask-Bootstrap4==4.0.2
dependency-injector==4.40.0
pytest==7.1.3
tensorflow==2.10.0
keras==2.10.0
boto3==1.24.92